src/backend/file_checker/container/requirements.txt
python-libmagic == 0.4.0